Rename WelcomePanelScrollViewGradient > ScrollViewGradientView
