+++ Wikipedia/Code/WelcomePanelViewController.xib
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="22155" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22131"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
@@ -73,7 +71,7 @@
                                         <constraint firstItem="UVD-iW-SNC" firstAttribute="width" secondItem="1on-Ec-EaB" secondAttribute="width" id="Yak-sj-SIx"/>
                                     </constraints>
                                 </scrollView>
-                                <view opaque="NO" clipsSubviews="YES" clearsContextBeforeDrawing="NO" userInteractionEnabled="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Yah-7B-ltZ" customClass="WelcomePanelScrollViewGradient" customModule="Wikipedia" customModuleProvider="target">
+                                <view opaque="NO" clipsSubviews="YES" clearsContextBeforeDrawing="NO" userInteractionEnabled="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Yah-7B-ltZ" customClass="ScrollViewGradientView" customModule="Wikipedia" customModuleProvider="target">
                                     <rect key="frame" x="0.0" y="0.0" width="355" height="541"/>
                                     <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 </view>
@@ -104,7 +102,7 @@
                                                 <fontDescription key="fontDescription" style="UICTFontTextStyleSubhead"/>
                                             </variation>
                                         </label>
-                                        <button opaque="NO" contentMode="scaleToFill" verticalCompressionResistancePriority="1000" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="urh-zw-u6D" customClass="AutoLayoutSafeMultiLineButton" customModule="Wikipedia" customModuleProvider="target">
+                                        <button opaque="NO" contentMode="scaleToFill" verticalCompressionResistancePriority="1000" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="urh-zw-u6D" customClass="AutoLayoutSafeMultiLineButton" customModule="Wikipedia" customModuleProvider="target">
                                             <rect key="frame" x="0.0" y="36" width="327" height="50"/>
                                             <color key="backgroundColor" white="0.0" alpha="1" colorSpace="calibratedWhite"/>
                                             <constraints>
@@ -139,14 +137,20 @@
                     </subviews>
                 </stackView>
             </subviews>
+            <viewLayoutGuide key="safeArea" id="QJu-Hj-Iei"/>
             <constraints>
                 <constraint firstItem="wlh-UI-800" firstAttribute="top" secondItem="uba-xZ-kBn" secondAttribute="topMargin" constant="10" id="NBg-fo-a5o"/>
                 <constraint firstItem="wlh-UI-800" firstAttribute="leading" secondItem="uba-xZ-kBn" secondAttribute="leading" constant="10" id="d2x-iB-a3Y"/>
                 <constraint firstAttribute="trailing" secondItem="wlh-UI-800" secondAttribute="trailing" constant="10" id="iMn-E9-8tE"/>
                 <constraint firstAttribute="bottom" secondItem="wlh-UI-800" secondAttribute="bottom" id="ovg-X0-YVT"/>
             </constraints>
-            <viewLayoutGuide key="safeArea" id="QJu-Hj-Iei"/>
+            <point key="canvasLocation" x="142" y="19"/>
         </view>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
     </objects>
+    <designables>
+        <designable name="urh-zw-u6D">
+            <size key="intrinsicContentSize" width="51" height="33"/>
+        </designable>
+    </designables>
 </document>
